requirements.txt
amqp==2.6.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==1.16.0
eventlet==0.30.2
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.14
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==4.6.11
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.2
mock==4.0.3
nameko==2.13.0
path==15.1.2
path.py==12.5.0
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
urllib3==1.26.3
vine==1.3.0
Werkzeug==1.0.1
wrapt==1.12.1